fix(db): Cambiar tipo de dato de storage a decimal

+++ database/migrations/2025_02_13_173108_create_storage_devices_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string("image");
 
             $table->string("type");
-            $table->integer("storage");
+            $table->decimal("storage", 10, 2);
             $table->timestamps();
         });
     }
